A branch

+++ CapstonProjectBE/Controllers/PlatformController.cs
@@ -127,11 +127,11 @@ namespace CapstonProjectBE.Controllers
         public async Task<IActionResult> CreateProjectPlatform([FromForm]ProjectPlatformDTO projectPlatformDTO)
         {
             var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
-            //var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
-            //if (check != null)
-            //{
-            //    return check;
-            //}
+            var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
+            if (check != null)
+            {
+                return check;
+            }
             var result = await _platformService.CreateProjectPlatform(projectPlatformDTO);
             if (!result.Success)
             {
@@ -143,14 +143,14 @@ namespace CapstonProjectBE.Controllers
 
         [Authorize(Roles = "STAFF, CUSTOMER")]
         [HttpDelete("project/delete")]
-        public async Task<IActionResult> RemoveProjectPlatform(ProjectPlatformDTO projectPlatformDTO)
+        public async Task<IActionResult> RemoveProjectPlatform([FromForm] ProjectPlatformDTO projectPlatformDTO)
         {
             var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
-            //var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
-            //if (check != null)
-            //{
-            //    return check;
-            //}
+            var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
+            if (check != null)
+            {
+                return check;
+            }
             var result = await _platformService.RemoveProjectPlatform(projectPlatformDTO);
             if (!result.Success)
             {
